moving search query generator
